# chore: split lint and format