--- a/README.md
+++ b/README.md
@@ -28,7 +28,6 @@
 
 ## Table of Contents
 
-- [Table of Contents](#table-of-contents)
 - [Getting Started](#getting-started)
 - [Running from command line](#running-from-command-line)
 - [Additional Configuration](#additional-configuration)
@@ -37,7 +36,7 @@
   - [Using webpack](#using-webpack)
   - [Using Vite](#using-vite)
   - [Using Parcel](#using-parcel)
-  - [Using TypeScript](#using-typescript)
+  - [Using Typescript](#using-typescript)
 - [Documentation](#documentation)
 - [Badge](#badge)
 - [Contributing](#contributing)