<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

<!-- Please remember to update CHANGELOG.md at the root of the project if you have not done so. -->

## Summary

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

It's bad to run Prettier as ESLint plugin, it's super slow.

- [Why eslint-plugin-prettier Is Often Harmful](https://www.joshuakgoldberg.com/blog/you-probably-dont-need-eslint-config-prettier-or-eslint-plugin-prettier/#why-eslint-plugin-prettier-is-often-harmful)

Run eslint with `TIMING=1`

Before

```text
yarn eslint .
Rule                             | Time (ms) | Relative
:--------------------------------|----------:|--------:
prettier/prettier                | 22804.028 |    36.2%
import/default                   |  3602.487 |     5.7%
import/no-duplicates             |  3150.643 |     5.0%
unicorn/no-unnecessary-polyfills |  2926.342 |     4.6%
unicorn/expiring-todo-comments   |  2317.099 |     3.7%
import/namespace                 |  2275.440 |     3.6%
jsdoc/check-alignment            |  1874.229 |     3.0%
unicorn/no-thenable              |  1396.356 |     2.2%
import/order                     |   855.491 |     1.4%
unicorn/new-for-builtins         |   727.296 |     1.2%
```

After

```text
yarn eslint .
Rule                              | Time (ms) | Relative
:---------------------------------|----------:|--------:
import/default                    |  3129.741 |     8.6%
import/no-duplicates              |  2724.314 |     7.5%
unicorn/no-unnecessary-polyfills  |  2723.796 |     7.5%
unicorn/expiring-todo-comments    |  2009.008 |     5.5%
import/namespace                  |  1967.862 |     5.4%
jsdoc/check-alignment             |  1823.156 |     5.0%
unicorn/no-thenable               |  1150.524 |     3.2%
import/order                      |   794.733 |     2.2%
import/no-extraneous-dependencies |   658.272 |     1.8%
unicorn/prefer-array-flat         |   650.011 |     1.8%
```

Whole lint and format

Before

```
$ time yarn lint

real    1m39.628s
user    2m30.322s
sys     0m7.407s

$ time yarn lint:prettier . -w > /dev/null

real    0m27.648s
user    0m31.158s
sys     0m2.721s
```

After

```
$ time yarn lint

real    0m6.488s
user    0m6.935s
sys     0m1.005s

$ time yarn lint:prettier > /dev/null

real    0m14.757s
user    0m13.220s
sys     0m2.986s
```

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->

N/A