@SimenB 2025-04-18
In general I agree, but just having a single "format on save" which actually works is super nice 😅 especially since eslint autofixes can create style violations.

And at least on my beery mac, I don't really feel the perf hit